pyproject.toml
[project]
name = "ai-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Thielson Almendra",email = "thielson12@gmail.com"}
]
readme = "README.md"
requires-python = "<4.0,>=3.13"
packages = [{ include = "**/*.py", from = "." }]
dependencies = [
    "langchain (>=0.3.15,<0.4.0)",
    "openai (>=1.60.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "langchain-groq (>=0.2.3,<0.3.0)",
    "langchain-community (>=0.3.15,<0.4.0)",
    "langgraph (>=0.2.67,<0.3.0)",
    "langchain-core (>=0.3.31,<0.4.0)",
    "langchain-ollama (>=0.2.2,<0.3.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
chat = "ai_assistant.chat:chat_loop"
graph = "ai_assistant.graph_image:get_graph"